lklfuse: update to libfuse3
Update to the new libfuse3 API, which has a number of attractive
features for lkl:
- an lseek() hook with SEEK_DATA / SEEK_HOLE
- copy_file_range()
- RENAME_NOREPLACE and RENAME_EXCHANGE.

Aside from passing through the rename flags, the remaining features
still need to be added to lklfuse. This commit attempts to be a direct
switch over to the new API.

The last upstream release of libfuse2 was in 2019 (2.9.9), so it makes
sense to upgrade regardless of the new functionality.

Signed-off-by: David Disseldorp <ddiss@suse.de>
(cherry picked from commit fc0d27b8b37c089352a47d60ad44829a2ad63a4e)